CPlayer::BuildOffscreenSpideySenseIndicatorList
Decompile CPlayer::BuildOffscreenSpideySenseIndicatorList (0x4C5250, 480
bytes). Scans BaddyList for offscreen threats and builds the spidey-sense
indicator list (direction + handle per slot, min/max distance tracked in
new CPlayer fields).

Also fixes CBody::mPlayerDist from u16 to u32 (ob.h). The disassembly
always reads/compares it with a full 32-bit register and unsigned jcc's,
never a 16-bit op. Rebuilt and reverified CPowerUp::AI and
CPowerUp::CheckAge (powerup.cpp, the only other users) with cmpsum: still
0 mnemonic diffs after the change.

New CPlayer fields carved out of existing padding: field_8BC/field_8C0
(running max/min CBody::mPlayerDist this pass) and field_EC0 (found-any
flag). New SIndicator field: mDirection (CVector), the normalized
local-space direction stored per indicator slot, carved out of its old
leading PADDING(0xC).

13 mnemonic diffs left, all pure register-scheduling residue (same
instruction count and byte length as the original). Tagged
@Ok @AlmostMatching with the residue and 14 tried hypotheses logged in
the function comment.

+++ ob.h
@@ -150,7 +150,12 @@ public:
 	u16 mCollision;
 	i16 mHealth;
 
-	u16 mPlayerDist;
+	// @FIXME was u16; the disassembly always reads/compares this with a
+	// full 32-bit register (no movzx, no 16-bit ops) and unsigned jcc's
+	// (jbe/ja against positive constants in powerup.cpp), so the real
+	// field is u32. offset unchanged (0xE4, still ends exactly at
+	// field_E8's 0xE8).
+	u32 mPlayerDist;
 
 	CVector field_E8;
